+++ example/src/config.ts
@@ -1,0 +1,6 @@
+import { CachedFlagsConfig } from 'react-native-cached-flags';
+
+export const myConfig: CachedFlagsConfig = {
+  aspectRatio: '4:3',
+  useFallbackEmoji: true,
+};